requirement.txt
flask==1.0.2
flask-restful==0.3.6
requests==2.19.1
twitter==1.18.0
pytest-flask==0.13.0
mock==2.0.0
pytest-mock==1.10.0

